add support for a bunny.net img invalidator
