+++ packages/aws/src/index.ts
@@ -1,3 +1,5 @@
 export * from './kms'
 export * from './s3'
 export * from './cloudfront'
+export * from './bunny'
+export * from './types'
