+++ services/bsky/indexer.js
@@ -3,7 +3,7 @@
 require('dd-trace/init') // Only works with commonjs
 
 // Tracer code above must come before anything else
-const { CloudfrontInvalidator } = require('@atproto/aws')
+const { CloudfrontInvalidator, BunnyInvalidator } = require('@atproto/aws')
 const {
   IndexerConfig,
   BskyIndexer,
@@ -25,12 +25,18 @@ const main = async () => {
     dbPostgresUrl: env.dbPostgresUrl,
     dbPostgresSchema: env.dbPostgresSchema,
   })
-  const cfInvalidator = env.cfDistributionId
-    ? new CloudfrontInvalidator({
-        distributionId: env.cfDistributionId,
-        pathPrefix: cfg.imgUriEndpoint && new URL(cfg.imgUriEndpoint).pathname,
-      })
-    : undefined
+  let imgInvalidator
+  if (env.bunnyAccessKey) {
+    imgInvalidator = new BunnyInvalidator({
+      accessKey: env.bunnyAccessKey,
+      urlPrefix: cfg.imgUriEndpoint,
+    })
+  } else if (env.cfDistributionId) {
+    imgInvalidator = new CloudfrontInvalidator({
+      distributionId: env.cfDistributionId,
+      pathPrefix: cfg.imgUriEndpoint && new URL(cfg.imgUriEndpoint).pathname,
+    })
+  }
   const redis = new Redis(
     cfg.redisSentinelName
       ? {
@@ -47,7 +53,7 @@ const main = async () => {
     db,
     redis,
     cfg,
-    imgInvalidator: cfInvalidator,
+    imgInvalidator,
   })
   await indexer.start()
   process.on('SIGTERM', async () => {
@@ -77,6 +83,7 @@ const getEnv = () => ({
   dbPoolSize: maybeParseInt(process.env.DB_POOL_SIZE),
   dbPoolMaxUses: maybeParseInt(process.env.DB_POOL_MAX_USES),
   dbPoolIdleTimeoutMs: maybeParseInt(process.env.DB_POOL_IDLE_TIMEOUT_MS),
+  bunnyAccessKey: process.env.BUNNY_ACCESS_KEY,
   cfDistributionId: process.env.CF_DISTRIBUTION_ID,
   imgUriEndpoint: process.env.IMG_URI_ENDPOINT,
 })
